Organize RHSPlib the same way it's organized upstream
