+++ RHSPlib/RHSPlib_compiler.h
