+++ RHSPlib/RHSPlib_device_control.h
