+++ RHSPlib/RHSPlib_dio.c
