+++ RHSPlib/RHSPlib_errors.h
