+++ RHSPlib/RHSPlib_motor.c
