+++ RHSPlib/RHSPlib_motor.h
